+++ src/listas.html
@@ -53,6 +53,6 @@
             </div>
         </main>
         <audio id="intro1" src="../assets/audio/bcs.mp3"></audio>
-       <script src="app.js"></script>
+       <script src="listas.js"></script>
       </body>
     </html>
